pov adjust setting bug fix (screen dimensions were not integers)

+++ screen_info.py
@@ -27,8 +27,8 @@ except FileNotFoundError:
 if pov_condition == 2:
     if global_monitor_width / 16 <= global_monitor_height / 9:
         fullscreen_scale = round(global_monitor_width / global_swidth)
-        swidth = global_monitor_width / fullscreen_scale
-        sheight = swidth / 16 * 9
+        swidth = int(global_monitor_width / fullscreen_scale)
+        sheight = int(swidth / 16 * 9)
 
 wiheight = global_sheight
 wiwidth = global_swidth
